+++ demo/demo.js
@@ -6,32 +6,17 @@ import Marker from 'pigeon-marker'
 import pigeonSvg from './incubator/pigeon.svg'
 import DraggableOverlay from './incubator/draggable-overlay'
 
-const mapboxEnabled = false
-
-// please change this if you take some code from here.
-// otherwise the demo page will run out of credits and that would be very sad :(
-const MAPBOX_ACCESS_TOKEN = 'pk.eyJ1IjoicGlnZW9uLW1hcHMiLCJhIjoiY2l3eW01Y2E2MDA4dzJ6cWh5dG9pYWlwdiJ9.cvdCf-7PymM1Y3xp5j71NQ'
-
-const mapbox = (mapboxId, accessToken) => (x, y, z, dpr) => {
-  return `https://api.mapbox.com/styles/v1/mapbox/${mapboxId}/tiles/256/${z}/${x}/${y}${dpr >= 2 ? '@2x' : ''}?access_token=${accessToken}`
-}
-
 const providers = {
   osm: (x, y, z) => {
     const s = String.fromCharCode(97 + (x + y + z) % 3)
     return `https://${s}.tile.openstreetmap.org/${z}/${x}/${y}.png`
   },
-  wikimedia: (x, y, z, dpr) => {
-    return `https://maps.wikimedia.org/osm-intl/${z}/${x}/${y}${dpr >= 2 ? '@2x' : ''}.png`
-  },
-  stamen: (x, y, z, dpr) => {
+  stamenTerrain: (x, y, z, dpr) => {
     return `https://stamen-tiles.a.ssl.fastly.net/terrain/${z}/${x}/${y}${dpr >= 2 ? '@2x' : ''}.jpg`
   },
-  streets: mapbox('streets-v10', MAPBOX_ACCESS_TOKEN),
-  satellite: mapbox('satellite-streets-v10', MAPBOX_ACCESS_TOKEN),
-  outdoors: mapbox('outdoors-v10', MAPBOX_ACCESS_TOKEN),
-  light: mapbox('light-v9', MAPBOX_ACCESS_TOKEN),
-  dark: mapbox('dark-v9', MAPBOX_ACCESS_TOKEN)
+  stamenToner: (x, y, z, dpr) => {
+    return `https://stamen-tiles.a.ssl.fastly.net/toner/${z}/${x}/${y}${dpr >= 2 ? '@2x' : ''}.png`
+  }
 }
 
 const markers = {
@@ -51,27 +36,9 @@ const Banner = () => (
   </a>
 )
 
-function isMapBox (provider) {
-  return provider === 'streets' || provider === 'satellite' || provider === 'outdoors' || provider === 'light' || provider === 'dark'
-}
-
-const MapboxAttribution = () => (
-  <span className='map-attribution'>
-    <span>© <a href='https://www.mapbox.com/about/maps/'>Mapbox</a></span>{' | '}
-    <span>© <a href='http://www.openstreetmap.org/copyright'>OpenStreetMap</a></span>{' | '}
-    <strong><a href='https://www.mapbox.com/map-feedback/' target='_blank'>Improve this map</a></strong>
-  </span>
-)
-
 const StamenAttribution = () => (
   <span className='map-attribution'>
     Map tiles by <a href="http://stamen.com">Stamen Design</a>, under <a href="http://creativecommons.org/licenses/by/3.0">CC BY 3.0</a>. Data by <a href="http://openstreetmap.org">OpenStreetMap</a>, under <a href="http://www.openstreetmap.org/copyright">ODbL</a>.
-  </span>
-)
-
-const WikimediaAttribution = () => (
-  <span className='map-attribution'>
-    Map tiles by <a href='https://foundation.wikimedia.org/w/index.php?title=Maps_Terms_of_Use#Where_does_the_map_data_come_from.3F'>Wikimedia</a>. Data by <a href="http://openstreetmap.org">OpenStreetMap</a>
   </span>
 )
 
@@ -82,7 +49,7 @@ export default class App extends Component {
     this.state = {
       center: [50.1102, 3.1506],
       zoom: 6,
-      provider: 'wikimedia',
+      provider: 'osm',
       metaWheelZoom: false,
       twoFingerDrag: false,
       animate: true,
@@ -157,13 +124,11 @@ export default class App extends Component {
             minZoom={minZoom}
             maxZoom={maxZoom}
             attribution={
-              isMapBox(provider)
-                ? <MapboxAttribution />
-                : provider === 'stamen'
-                  ? <StamenAttribution />
-                  : provider === 'wikimedia'
-                    ? <WikimediaAttribution />
-                    : null}
+              provider === 'stamenTerrain' || provider === 'stamenToner'
+                ? <StamenAttribution />
+                : provider === 'wikimedia'
+                  ? <WikimediaAttribution />
+                  : null}
             defaultWidth={600}
             height={400}
             boxClassname="pigeon-filters">
@@ -181,7 +146,6 @@ export default class App extends Component {
                 width={100}
                 height={95} />
             </DraggableOverlay>
-            {isMapBox(provider) && <span className='mapbox-wordmark' />}
           </Map>
         </div>
         <div>
@@ -200,8 +164,8 @@ export default class App extends Component {
           {Object.keys(providers).map(key => (
             <button
               key={key}
-              onClick={() => isMapBox(key) && !mapboxEnabled ? window.alert('Mapbox tiles disabled! See issue #33 for details!') : this.setState({ provider: key })}
-              style={{fontWeight: provider === key ? 'bold' : 'normal', color: isMapBox(key) && !mapboxEnabled ? '#aaa' : '#000'}}>
+              onClick={() => this.setState({ provider: key })}
+              style={{fontWeight: provider === key ? 'bold' : 'normal'}}>
               {key}
             </button>
           ))}