+++ CHANGELOG.md
@@ -7,6 +7,12 @@ Once we reach 1.0 all deprecations will be removed and the project will switch t
 
 ## Uncommitted
 
+## 0.14.0 - 2020-03-03
+### Changes
+- Changed the default tile provider from the broken Wikimedia to OSM
+- Add more information about setting up your own tile provider
+- Add lazy loading for tiles (#87 by @maxsteenbergen)
+
 ## 0.13.0 - 2019-05-09
 ### Changes
 - Add the `dprs` parameter to `<Map />` and `dpr` as the 4th argument for the `provider` functions.